split key, packed, rework packed
